Order Detail Control
- Order 화면은 Favorite Menu View와 Order Detail View로 구성되도록 수정
- 기본적으로 xd의 흐름과는 같게 이동할 수 있도록 수정
- Navigation View와 Navigation Link 사용 시도 중
